requirements.txt
requests>=2.20.0
PyYAML>=5.3.1
fastapi>=0.85.0
uvicorn[standard]>=0.18.0
pydantic>=1.10.2
httpx>=0.24.0


